Silly Python version bug

+++ keras/force_pygc.m
@@ -4,7 +4,10 @@ function num_collected = force_pygc()
 % Matlab jobs which use the py.* interface can leak memory badly (I found
 % this was happening to GPU memory when I was using Keras).
 gc = py.importlib.import_module('gc');
-num_collected = gc.collect.feval();
+% gc.collect.feval() returns an integer. For some reason, Matlab only
+% converts that to a Matlab-native integer (as opposed to a py.int)
+% *sometimes*. The int64() wrap should work fine.
+num_collected = int64(gc.collect.feval());
 if num_collected > 0
     fprintf('Python collected %i garbage objects\n', num_collected);
 end
